requirements.txt
-e git+https://github.com/Tivix/django-common.git@bcf5a67192f967a71688f597dc79db8f9b620412#egg=django_common_helpers-master
Babel==2.2.0
Django==1.9.7
Pillow==3.0.0
PyJWT==1.4.0
Pygments==2.0.2
WebOb==1.4.1
WebTest==2.0.18
Werkzeug==0.10.4
appnope==0.1.0
beautifulsoup4==4.4.1
blessings==1.6
bpython==0.14.2
click==6.2
contextlib2==0.5.3
curtsies==0.1.19
decorator==4.0.4
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.23.0
django-braces==1.9.0
django-common==0.1.51
django-cors-headers==1.1.0
django-extensions==1.6.1
django-filter==0.11.0
django-oauth-toolkit==0.10.0
django-oauth2-provider==0.2.6.1
django-phonenumber-field==1.0.0
django-push-notifications==1.4.1
django-rest-auth==0.7.0
django-rest-framework-social-oauth2==1.0.4
django-toolbelt==0.0.1
django-twilio-sms==0.2
django-twilio==0.8.0
django-vanilla-views==1.0.4
django-webtest==1.7.8
django-widget-tweaks==1.4.1
djangorestframework-filters==0.6.0
djangorestframework==3.3.3
first==2.0.1
geopy==1.11.0
#gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.3.0
httplib2==0.9.2
ipdb==0.8.1
ipython-genutils==0.1.0
ipython==4.0.0
mock==1.0.0
model-mommy==1.2.5
oauth2-provider==0.0
oauthlib==1.1.2
path.py==8.1.1
#pexpect==4.1.0
phonenumbers==7.2.3
pickleshare==0.5
pip-tools==1.4.2
psycopg2==2.6.1
pydot==1.0.2
pygraphviz==1.3.1
pyparsing==1.5.7
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2015.7
redis==2.10.5
requests-oauthlib==0.6.1
requests==2.10.0
shortuuid==0.4.3
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
tinycss==0.3
tornado==4.3
tornado-redis=2.4.18
traitlets==4.0.0
twilio==4.9.2
waitress==0.8.10
